alzemiro changes

+++ application/applicationGenerator.sh
@@ -21,7 +21,12 @@ do
 			cp -u -v source/applicationGeneric/applicationGeneric.c application$i.c
 		fi
 	else
-		cp -u -v source/applicationGeneric/applicationGeneric.c application$i.c
+		file="source/migration_scenario/migration_scenario_main.c"
+		if [ -f "$file" ]; then
+			cp -u -v source/migration_scenario/migration_scenario_main.c application$i.c
+		else 
+			cp -u -v source/applicationGeneric/applicationGeneric.c application$i.c
+		fi
 	fi
 	
 done
